refactoring, part3: pytest classes extracted
