+++ app_build_suite/__main__.py
@@ -12,7 +12,7 @@ from app_build_suite.build_steps import (
 )
 from app_build_suite.build_steps.build_step import STEP_ALL
 from app_build_suite.errors import ConfigError
-from app_build_suite.build_steps.base_test_runner import PytestTestFilteringPipeline
+from app_build_suite.build_steps.pytest import PytestTestFilteringPipeline
 from .components import ComponentsContainer, Runner
 
 ver = "0.0.1-dev"
